loginRegister

+++ database/migrations/2026_07_06_041949_add_payment_bill_code_to_orders_table.php
@@ -10,16 +10,20 @@ return new class extends Migration
      * Run the migrations.
      */
     public function up()
-        {
+    {
+        if (!Schema::hasColumn('orders', 'payment_bill_code')) {
             Schema::table('orders', function (Blueprint $col) {
-                $col->string('payment_bill_code')->nullable()->after('id'); // letak lepas id atau mana-mana kolum sesuai
+                $col->string('payment_bill_code')->nullable()->after('id');
             });
         }
+    }
 
     public function down()
-        {
+    {
+        if (Schema::hasColumn('orders', 'payment_bill_code')) {
             Schema::table('orders', function (Blueprint $col) {
                 $col->dropColumn('payment_bill_code');
             });
         }
+    }
 };
